container/requirements.txt
tensorflow==2.4.1
keras==2.4.3
numpy==1.19.5
pillow
flask
gunicorn